+++ packages/salesforcedx-utils-vscode/src/telemetry/index.ts
@@ -5,6 +5,6 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { TelemetryService } from './telemetry';
+import { TelemetryService } from '../services/telemetry';
 
 export const telemetryService = TelemetryService.getInstance();
